+++ src/mdurocherart/models.py
@@ -1,7 +1,9 @@
 from pydantic import BaseModel, EmailStr, Field
+from typing import Optional
 
 
 class Email(BaseModel):
     email: EmailStr = Field(...)
     subject: str = Field(...)
     body: str = Field(...)
+    name: Optional[str]
